sub-divided the artifact testing into a a seperate file for each CRUD component. This makes it make readable, easier to use and contribute. The previous single file was over 1000 lines and a mess. Now it is much more approachable
